# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API de Donaciones - Registro de donantes
  version: 1.0.0
  description: Módulo de registro de personas donantes con autenticación JWT y roles (admin / usuario).
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
    Credentials:
      type: object
      required: [email, password]
      properties:
        email: { type: string, example: admin@donaciones.local }
        password: { type: string, example: Admin123! }
    Register:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, example: Ana López }
        email: { type: string, example: ana@correo.com }
        password: { type: string, example: Segura123 }
    Donor:
      type: object
      required: [name, email, type]
      properties:
        name: { type: string, example: Supermercado La Esperanza }
        email: { type: string, example: contacto@esperanza.mx }
        phone: { type: string, example: '+52 81 1234 5678' }
        type: { type: string, enum: [persona, empresa] }
        address: { type: string, example: Av. Constitución 100 }
        notes: { type: string, example: Dona pan los viernes }
paths:
  /health:
    get: { summary: Estado del servicio, responses: { '200': { description: OK } } }
  /api/auth/register:
    post:
      summary: Registrar usuario (rol usuario)
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Register' } } } }
      responses: { '201': { description: Creado }, '400': { description: Datos inválidos }, '409': { description: Duplicado } }
  /api/auth/login:
    post:
      summary: Iniciar sesión y obtener JWT
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Credentials' } } } }
      responses: { '200': { description: OK }, '401': { description: Credenciales inválidas } }
  /api/auth/me:
    get:
      summary: Usuario autenticado
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: OK }, '401': { description: No autenticado } }
  /api/donors:
    get:
      summary: Listar donantes (admin ve todos, usuario solo los suyos)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string, enum: [persona, empresa] } }
      responses: { '200': { description: OK } }
    post:
      summary: Registrar donante
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Donor' } } } }
      responses: { '201': { description: Creado }, '400': { description: Datos inválidos } }
  /api/donors/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      summary: Detalle de donante
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: OK }, '403': { description: Sin acceso }, '404': { description: No existe } }
    put:
      summary: Actualizar donante
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Donor' } } } }
      responses: { '200': { description: OK } }
    delete:
      summary: Eliminar donante (solo admin)
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Eliminado }, '403': { description: Solo admin } }
  /api/users:
    get:
      summary: Listar usuarios (solo admin)
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: OK }, '403': { description: Solo admin } }
  /api/users/{id}/role:
    patch:
      summary: Cambiar rol (solo admin)
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, properties: { role: { type: string, enum: [admin, usuario] } } } } } }
      responses: { '200': { description: OK } }